requirements.txt
python-fasthtml
monsterui
fh-posts